+++ pkg/analyzer/lib/src/context/builder.dart
@@ -5,7 +5,6 @@
 import 'dart:collection';
 import 'dart:core';
 
-import 'package:analyzer/dart/analysis/context_locator.dart' as api;
 import 'package:analyzer/dart/analysis/declared_variables.dart';
 import 'package:analyzer/file_system/file_system.dart';
 import 'package:analyzer/src/analysis_options/analysis_options_provider.dart';
@@ -14,6 +13,7 @@ import 'package:analyzer/src/command_line/arguments.dart'
 import 'package:analyzer/src/context/context_root.dart';
 import 'package:analyzer/src/context/packages.dart';
 import 'package:analyzer/src/dart/analysis/byte_store.dart';
+import 'package:analyzer/src/dart/analysis/context_locator.dart';
 import 'package:analyzer/src/dart/analysis/driver.dart'
     show AnalysisDriver, AnalysisDriverScheduler;
 import 'package:analyzer/src/dart/analysis/driver_based_analysis_context.dart'
@@ -110,9 +110,10 @@ class ContextBuilder {
 
   /// Return an analysis driver that is configured correctly to analyze code in
   /// the directory with the given [path].
-  AnalysisDriver buildDriver(ContextRoot contextRoot) {
+  AnalysisDriver buildDriver(ContextRoot contextRoot, Workspace workspace) {
     String path = contextRoot.root;
-    var options = getAnalysisOptions(path, contextRoot: contextRoot);
+
+    var options = getAnalysisOptions(path, workspace, contextRoot: contextRoot);
     //_processAnalysisOptions(context, optionMap);
 
     SummaryDataStore? summaryData;
@@ -121,12 +122,6 @@ class ContextBuilder {
       summaryData = SummaryDataStore(librarySummaryPaths);
     }
 
-    Workspace workspace = ContextBuilder.createWorkspace(
-      resourceProvider: resourceProvider,
-      options: builderOptions,
-      rootPath: path,
-      lookForBazelBuildFileSubstitutes: lookForBazelBuildFileSubstitutes,
-    );
     final sf =
         createSourceFactoryFromWorkspace(workspace, summaryData: summaryData);
 
@@ -150,18 +145,18 @@ class ContextBuilder {
     );
 
     // Set API AnalysisContext for the driver.
-    var apiContextRoots = api.ContextLocator(
+    var apiContextRoots = ContextLocatorImpl(
       resourceProvider: resourceProvider,
     ).locateRoots(
       includedPaths: [contextRoot.root],
       excludedPaths: contextRoot.exclude,
+      overrideWorkspace: workspace,
     );
     driver.configure(
       analysisContext: api.DriverBasedAnalysisContext(
         resourceProvider,
         apiContextRoots.first,
         driver,
-        workspace: workspace,
       ),
     );
 
@@ -206,13 +201,8 @@ class ContextBuilder {
 //    }
 //  }
 
-  SourceFactory createSourceFactory(String rootPath,
+  SourceFactory createSourceFactory(String rootPath, Workspace workspace,
       {SummaryDataStore? summaryData}) {
-    Workspace workspace = ContextBuilder.createWorkspace(
-      resourceProvider: resourceProvider,
-      options: builderOptions,
-      rootPath: rootPath,
-    );
     DartSdk sdk = findSdk(workspace);
     if (summaryData != null && sdk is SummaryBasedDartSdk) {
       summaryData.addBundle(null, sdk.bundle);
@@ -288,7 +278,7 @@ class ContextBuilder {
   /// Return the analysis options that should be used to analyze code in the
   /// directory with the given [path]. Use [verbosePrint] to echo verbose
   /// information about the analysis options selection process.
-  AnalysisOptionsImpl getAnalysisOptions(String path,
+  AnalysisOptionsImpl getAnalysisOptions(String path, Workspace workspace,
       {void Function(String text)? verbosePrint, ContextRoot? contextRoot}) {
     void verbose(String text) {
       if (verbosePrint != null) {
@@ -296,13 +286,6 @@ class ContextBuilder {
       }
     }
 
-    // TODO(danrubel) restructure so that we don't create a workspace
-    // both here and in createSourceFactory
-    Workspace workspace = ContextBuilder.createWorkspace(
-      resourceProvider: resourceProvider,
-      options: builderOptions,
-      rootPath: path,
-    );
     SourceFactory sourceFactory = workspace.createSourceFactory(null, null);
     AnalysisOptionsProvider optionsProvider =
         AnalysisOptionsProvider(sourceFactory);
@@ -420,13 +403,19 @@ class ContextBuilder {
     }
   }
 
+  /// If [packages] is provided, it will be used for the [Workspace],
+  /// otherwise the packages file from [options] will be used, or discovered
+  /// from [rootPath].
+  ///
+  /// TODO(scheglov) Make [packages] required, remove [options] and discovery.
   static Workspace createWorkspace({
     required ResourceProvider resourceProvider,
     required ContextBuilderOptions options,
+    Packages? packages,
     required String rootPath,
     bool lookForBazelBuildFileSubstitutes = true,
   }) {
-    var packages = ContextBuilder.createPackageMap(
+    packages ??= ContextBuilder.createPackageMap(
       resourceProvider: resourceProvider,
       options: options,
       rootPath: rootPath,