+++ pkg/analyzer/lib/src/dart/analysis/driver_based_analysis_context.dart
@@ -6,7 +6,6 @@ import 'package:analyzer/dart/analysis/analysis_context.dart';
 import 'package:analyzer/dart/analysis/context_root.dart';
 import 'package:analyzer/dart/analysis/session.dart';
 import 'package:analyzer/file_system/file_system.dart';
-import 'package:analyzer/src/context/builder.dart';
 import 'package:analyzer/src/dart/analysis/driver.dart' show AnalysisDriver;
 import 'package:analyzer/src/dart/sdk/sdk.dart';
 import 'package:analyzer/src/generated/engine.dart' show AnalysisOptions;
@@ -23,16 +22,14 @@ class DriverBasedAnalysisContext implements AnalysisContext {
   /// The driver on which this context is based.
   final AnalysisDriver driver;
 
-  /// The [Workspace] for this context, `null` if not yet created.
-  Workspace? _workspace;
-
   /// Initialize a newly created context that uses the given [resourceProvider]
   /// to access the file system and that is based on the given analysis
   /// [driver].
   DriverBasedAnalysisContext(
-      this.resourceProvider, this.contextRoot, this.driver,
-      {Workspace? workspace})
-      : _workspace = workspace {
+    this.resourceProvider,
+    this.contextRoot,
+    this.driver,
+  ) {
     driver.analysisContext = this;
   }
 
@@ -51,17 +48,9 @@ class DriverBasedAnalysisContext implements AnalysisContext {
     return null;
   }
 
+  @Deprecated('Use contextRoot.workspace instead')
   @override
   Workspace get workspace {
-    return _workspace ??= _buildWorkspace();
-  }
-
-  Workspace _buildWorkspace() {
-    var path = contextRoot.root.path;
-    return ContextBuilder.createWorkspace(
-      resourceProvider: resourceProvider,
-      options: ContextBuilderOptions(),
-      rootPath: path,
-    );
+    return contextRoot.workspace;
   }
 }