+++ pkg/front_end/test/compile_with_no_sdk_test.dart
@@ -7,8 +7,7 @@
 import 'package:front_end/src/api_prototype/compiler_options.dart';
 import 'package:front_end/src/api_prototype/memory_file_system.dart';
 
-import 'incremental_load_from_dill_suite.dart'
-    show TestIncrementalCompiler, getOptions;
+import 'incremental_suite.dart' show TestIncrementalCompiler, getOptions;
 
 main() async {
   await compile("import 'foo.dart' if (dart.library.bar) 'baz.dart';");