+++ pkg/front_end/test/incremental_load_from_invalid_dill_test.dart
@@ -45,7 +45,7 @@ import 'package:front_end/src/fasta/kernel/utils.dart' show serializeComponent;
 import 'package:kernel/kernel.dart'
     show Component, Library, NonNullableByDefaultCompiledMode;
 
-import 'incremental_load_from_dill_suite.dart' show getOptions;
+import 'incremental_suite.dart' show getOptions;
 
 Future<Null> main() async {
   Tester tester = new Tester();